Fix parsing, fix socket closing after one message
